requirements.txt
django==4.2.0
python-dotenv
djangorestframework
psycopg2-binary